+++ src/app/src/components/FilterSidebar.jsx
@@ -35,6 +35,7 @@ import {
 import {
     contributorOptionsPropType,
     countryOptionsPropType,
+    sectorOptionsPropType,
 } from '../util/propTypes';
 
 import { allListsAreEmpty } from '../util/util';
@@ -58,6 +59,7 @@ class FilterSidebar extends Component {
         const {
             contributorsData,
             countriesData,
+            sectorsData,
             fetchFilterOptions,
             fetchContributors,
             fetchLists,
@@ -65,7 +67,7 @@ class FilterSidebar extends Component {
             contributors,
         } = this.props;
 
-        if (allListsAreEmpty(contributorsData, countriesData)) {
+        if (allListsAreEmpty(contributorsData, countriesData, sectorsData)) {
             return fetchFilterOptions();
         }
 
@@ -207,6 +209,7 @@ FilterSidebar.propTypes = {
     fetchCountries: func.isRequired,
     contributorsData: contributorOptionsPropType.isRequired,
     countriesData: countryOptionsPropType.isRequired,
+    sectorsData: sectorOptionsPropType.isRequired,
     vectorTileFeatureIsActive: bool.isRequired,
     fetchingFeatureFlags: bool.isRequired,
 };
@@ -217,6 +220,7 @@ function mapStateToProps({
         contributors: { data: contributorsData },
         lists: { data: listsData },
         countries: { data: countriesData },
+        sectors: { data: sectorsData },
     },
     featureFlags: { flags, fetching: fetchingFeatureFlags },
     embeddedMap: { embed },
@@ -229,6 +233,7 @@ function mapStateToProps({
         activeFilterSidebarTab,
         contributorsData,
         countriesData,
+        sectorsData,
         listsData,
         vectorTileFeatureIsActive: get(flags, 'vector_tile', false),
         fetchingFeatureFlags,